+++ src/vectors.hpp
@@ -8,10 +8,10 @@ class Vectors : public Drawable
 {
 	struct Vector
 	{
-		glm::vec3 pos;
-		glm::vec3 vector;
+		glm::dvec3 pos;
+		glm::dvec3 vector;
 		glm::vec3 color;
-		Vector(glm::vec3 pos, glm::vec3 vector, glm::vec3 color) : pos(pos), vector(vector), color(color) {}
+		Vector(glm::dvec3 pos, glm::dvec3 vector, glm::vec3 color) : pos(pos), vector(vector), color(color) {}
 	};
 
 	std::vector<Vector> vectors;
@@ -21,7 +21,7 @@ class Vectors : public Drawable
 public:
 	void init();
 
-	void addVector(glm::vec3 pos, glm::vec3 vector, glm::vec3 color);
+	void addVector(glm::dvec3 pos, glm::dvec3 vector, glm::vec3 color);
 
 	void clear();
